% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmmer.R
\name{hmmsearch}
\alias{hmmsearch}
\alias{hmmsearch,Hmmer-method}
\title{hmmsearch}
\usage{
hmmsearch(object, ...)
}
\arguments{
\item{object}{Hmmer object.}

\item{...}{further arguments passed to hmmsearch().}
}
\description{
hmmsearch
}
