% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmmer.R
\docType{methods}
\name{print,hmmer-method}
\alias{print,hmmer-method}
\title{print}
\usage{
\S4method{print}{hmmer}(x, ...)
}
\arguments{
\item{x}{hmmer object.}

\item{...}{further arguments (ignored).}
}
\description{
print
}
