% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmmer.R
\docType{data}
\name{hmmer}
\alias{hmmer}
\title{Initialize a HMMER docker image}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
hmmer$new()
}
\description{
Initializes a HMMER docker image and calls specific methods
}
\examples{
h <- hmmer$new()
h$image
h$outdir
h$hmmsearch("model.hmm", "sequences.fa")

}
\keyword{datasets}
