% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmmer.R
\docType{methods}
\name{print,Hmmer-method}
\alias{print,Hmmer-method}
\title{print}
\usage{
\S4method{print}{Hmmer}(x, ...)
}
\arguments{
\item{x}{Hmmer object.}

\item{...}{further arguments (ignored).}
}
\description{
print
}
